add changes in permissions

+++ car-racing-tournament-api/Controllers/SeasonController.cs
@@ -55,6 +55,9 @@ namespace car_racing_tournament_api.Controllers
         [HttpPut("{id}"), Authorize]
         public async Task<IActionResult> Put(Guid id, [FromBody] SeasonDto seasonDto)
         {
+            if (!await _userSeasonService.IsAdmin(new Guid(User.Identity!.Name!), id))
+                return Forbid();
+
             var resultUpdate = await _seasonService.UpdateSeason(id, seasonDto);
             if (!resultUpdate.IsSuccess)
                 return BadRequest(resultUpdate.ErrorMessage);
@@ -65,6 +68,9 @@ namespace car_racing_tournament_api.Controllers
         [HttpDelete("{id}"), Authorize]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!await _userSeasonService.IsAdmin(new Guid(User.Identity!.Name!), id))
+                return Forbid();
+
             var resultDelete = await _seasonService.DeleteSeason(id);
             if (!resultDelete.IsSuccess)
                 return BadRequest(resultDelete.ErrorMessage);
@@ -89,7 +95,7 @@ namespace car_racing_tournament_api.Controllers
         [HttpPost("{seasonId}/user-season"), Authorize]
         public async Task<IActionResult> Post(Guid seasonId, [FromForm] string usernameEmail)
         {
-            if (!await _userSeasonService.HasPermission(new Guid(User.Identity!.Name!), seasonId))
+            if (!await _userSeasonService.IsAdmin(new Guid(User.Identity!.Name!), seasonId))
                 return Forbid();
 
             var resultGet = await _userService.GetUserByUsernameEmail(usernameEmail);
@@ -116,7 +122,7 @@ namespace car_racing_tournament_api.Controllers
         [HttpPost("{seasonId}/driver"), Authorize]
         public async Task<IActionResult> PostDriver(Guid seasonId, [FromBody] DriverDto driverDto)
         {
-            if (!await _userSeasonService.HasPermission(new Guid(User.Identity!.Name!), seasonId))
+            if (!await _userSeasonService.IsAdminModerator(new Guid(User.Identity!.Name!), seasonId))
                 return Forbid();
 
             var resultAdd = await _seasonService.AddDriver(seasonId, driverDto);
@@ -139,7 +145,7 @@ namespace car_racing_tournament_api.Controllers
         [HttpPost("{seasonId}/team"), Authorize]
         public async Task<IActionResult> PostTeam(Guid seasonId, [FromBody] TeamDto team)
         {
-            if (!await _userSeasonService.HasPermission(new Guid(User.Identity!.Name!), seasonId))
+            if (!await _userSeasonService.IsAdminModerator(new Guid(User.Identity!.Name!), seasonId))
                 return Forbid();
 
             var resultAdd = await _seasonService.AddTeam(seasonId, team);
@@ -162,7 +168,7 @@ namespace car_racing_tournament_api.Controllers
         [HttpPost("{seasonId}/race"), Authorize]
         public async Task<IActionResult> PostRace(Guid seasonId, [FromBody] RaceDto raceDto)
         {
-            if (!await _userSeasonService.HasPermission(new Guid(User.Identity!.Name!), seasonId))
+            if (!await _userSeasonService.IsAdminModerator(new Guid(User.Identity!.Name!), seasonId))
                 return Forbid();
             
             var resultAdd = await _seasonService.AddRace(seasonId, raceDto);
